feat(web): extract entry normalize lib and expand normalization tests
Move registry entry normalization helpers into a pure library module
with focused unit coverage for platform inference, trust/source posture,
compatibility mapping, and related-entry assembly while keeping
contributor attribution logic in the thin data wrapper.
